pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "zlp"
version = "0.1.0"
description = "Single-workspace Zulip CLI toolkit for humans and AI agents — read, write, archive, and grep one Zulip workspace from the shell."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.9"
authors = [
    { name = "Jinguo Liu", email = "cacate0129@gmail.com" },
]
keywords = ["zulip", "cli", "chat", "archive", "agent", "ai", "llm", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "zulip>=0.9.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/GiggleLiu/zulip-management"
Issues = "https://github.com/GiggleLiu/zulip-management/issues"
Source = "https://github.com/GiggleLiu/zulip-management"

[project.scripts]
zlp = "zlp.cli:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/zlp"]

[tool.hatch.build.targets.sdist]
include = [
    "src/zlp",
    "tests",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.ruff]
line-length = 110
target-version = "py39"
